added hardware pulse-wave modulation to lirc_rpi
Amended lirc_rpi.c (kernel module) to support hardware pulse wave modulation. The new kernel module has an additional parameter that can be provided when loading the module: pwmcarrier. If pwmcarrier=1 (which is mutually exclusive with softcarrier=1) then hardware pulse-wave modulation is used. The GPIO out pin needs to be 18 as this is to my knowledge the only pwm-enabled pin.

Tested on RPi3 with raspbian / Linux 4.4.13-v7+
- with a DVD player (38 kHz, duty cycle 50%)
- with a visible light LED to confirm the effect of 25% (=dim), 50% and 75% (=bright) duty cycle.

The raspberry Pi is hard-coded to model 3 (by #define RPI 3). In theory, untested, model 2 utilizes the same memory addresses. The provided memory addresses for model 0 and 1 are also untested.